requirements.txt
saxo-openapi
python-dotenv
requests
flask
websocket-client
pytest
